Move CustomCheckboxView to shared file
